Fixed a unit test

+++ vol1/java-examples/src/main/java/com/heatonresearch/aifh/learning/score/ScoreClassificationData.java
@@ -68,7 +68,7 @@ public class ScoreClassificationData implements ScoreFunction {
             totalCount++;
             final int output = ralgo.computeClassification(aTrainingData.getInput());
 
-            if (output != aTrainingData.getIdeal()[0]) {
+            if (output != (int) aTrainingData.getIdeal()[0]) {
                 incorrectCount++;
             }
         }
